app/requirements.txt
Django
django-subcommands
djangorestframework
graphene-django
graphene-django-extras
jinja2
termcolor